SOLR-13125: add better tests

+++ solr/core/src/test/org/apache/solr/handler/component/TRAQueryComponentTest.java
@@ -19,20 +19,19 @@ package org.apache.solr.handler.component;
 
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.time.Instant;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
-import org.apache.commons.collections.iterators.ArrayIterator;
 import org.apache.lucene.util.IOUtils;
 import org.apache.solr.client.solrj.SolrRequest;
 import org.apache.solr.client.solrj.SolrServerException;
+import org.apache.solr.client.solrj.embedded.JettySolrRunner;
 import org.apache.solr.client.solrj.impl.CloudSolrClient;
 import org.apache.solr.client.solrj.request.CollectionAdminRequest;
 import org.apache.solr.client.solrj.request.ConfigSetAdminRequest;
@@ -42,15 +41,25 @@ import org.apache.solr.client.solrj.response.UpdateResponse;
 import org.apache.solr.cloud.SolrCloudTestCase;
 import org.apache.solr.cloud.api.collections.TimeRoutedAlias;
 import org.apache.solr.common.SolrInputDocument;
+import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.common.util.NamedList;
 import org.apache.solr.common.util.StrUtils;
+import org.apache.solr.core.SolrCore;
+import org.apache.solr.request.LocalSolrQueryRequest;
+import org.apache.solr.request.SolrQueryRequest;
+import org.apache.solr.response.SolrQueryResponse;
 import org.apache.solr.update.processor.TimeRoutedAliasUpdateProcessorTest;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.Test;
+import org.mockito.AdditionalAnswers;
+import org.mockito.Mockito;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 public class TRAQueryComponentTest extends SolrCloudTestCase {
 
@@ -69,6 +78,7 @@ public class TRAQueryComponentTest extends SolrCloudTestCase {
 
   @Before
   public void doBefore() throws Exception {
+    assumeWorkingMockito();
     configureCluster(4)
         .addConfig(optimizerConfName, configset(TRAConfigName))
         .configure();
@@ -133,20 +143,90 @@ public class TRAQueryComponentTest extends SolrCloudTestCase {
     )));
     solrClient.commit(alias);
 
-    // set debug to true to assert only right shards were queried
-    final QueryResponse qResp = solrClient.query(alias, params("q", "*:*", "sort", timeField + " desc",
+    // set debug to true to assert only right shards were queried, desc sort
+    QueryResponse qResp = solrClient.query(alias, params("q", "*:*", "sort", timeField + " desc",
         "rows", "2", "debug", "true")
     );
-    final String debugInfo = (String)qResp.getDebugMap().getOrDefault(RoutedAliasOptimizeQueryComponent.COMPONENT_NAME, "");
+    String debugInfo = getComponentDebugInfo(qResp);
     assertNotNull("debug info was not added", debugInfo);
     List<String> shards = getShardsFromQueryDebug(debugInfo);
-    assertTrue("was  expected to only get fields from '_2017-10-26', but instead got : " + shards,
-        shards.stream().anyMatch(x -> x.contains("_2017-10-26")));
     assertEquals(6, qResp.getResults().getNumFound());
     assertEquals(1, shards.size());
-    // assertEquals();
+    assertTrue("expected to only get fields from '_2017-10-26', but instead got : " + shards,
+        shards.stream().anyMatch(x -> x.contains("_2017-10-26")));
+
     assertEquals(Instant.parse("2017-10-26T07:00:00Z"), ((Date)qResp.getResults().get(0).get(timeField)).toInstant());
     assertEquals(Instant.parse("2017-10-26T06:00:00Z"), ((Date)qResp.getResults().get(1).get(timeField)).toInstant());
+
+    // set debug to true to assert only right shards were queried, asc sort
+    qResp = solrClient.query(alias, params("q", "*:*", "sort", timeField + " asc",
+        "rows", "2", "debug", "true")
+    );
+    debugInfo = getComponentDebugInfo(qResp);
+    assertNotNull("debug info was not added", debugInfo);
+    shards = getShardsFromQueryDebug(debugInfo);
+    assertEquals(6, qResp.getResults().getNumFound());
+    assertEquals(3, shards.size()); // two collections but one of them has 2 shards, so 3 in total
+    assertTrue("expected to only get fields from '" + alias + "_2017-10-23, '" + alias + "_2017-10-24', but instead got : " + String.join(", ", shards),
+        collectionMatchesRegex(shards, alias + ".*_2017-10-23_shard.*", alias + ".*_2017-10-24_shard.*"));
+  }
+
+  private String getComponentDebugInfo(QueryResponse qResp) {
+    return (String) qResp.getDebugMap().getOrDefault(RoutedAliasOptimizeQueryComponent.COMPONENT_NAME, "");
+  }
+
+  @Test
+  public void testComponentConditions() throws Exception {
+
+    createConfigSet(TRAConfigName, false);
+
+    cluster.waitForAllNodes(Integer.MAX_VALUE);
+
+    assertNotNull("jetty runner should be up", cluster.getJettySolrRunner(0));
+
+    RoutedAliasOptimizeQueryComponent component = new RoutedAliasOptimizeQueryComponent();
+    QueryComponent queryComponent = new QueryComponent();
+    // new SolrQueryReqMock(alias, new ModifiableSolrParams())
+    ResponseBuilder rb = createResponseBuilderWQuery(queryComponent, createQueryReq(alias, params()));
+
+    // should return fast since TRA does not have any collections.
+    assertEquals(ResponseBuilder.STAGE_DONE, component.distributedProcess(rb));
+
+    // delete collection
+    CollectionAdminRequest.deleteCollection(TRAConfigName).process(solrClient);
+
+    // create alias
+    CollectionAdminRequest.createTimeRoutedAlias(alias, "2017-10-23T00:00:00Z", "+1DAY", timeField,
+        CollectionAdminRequest.createCollection("_unused_", "tra-distrib", 1, 1)
+            .setMaxShardsPerNode(2))
+        .process(solrClient);
+
+    // no sort specified
+    rb = createResponseBuilderWQuery(queryComponent, createQueryReq(alias, params("q", "*:*")));
+    assertEquals(ResponseBuilder.STAGE_DONE, component.distributedProcess(rb));
+
+    // sort on field other than router.field
+    rb = createResponseBuilderWQuery(queryComponent, createQueryReq(alias, params("q", "*:*", "sort", "id asc")));
+    assertEquals(ResponseBuilder.STAGE_DONE, component.distributedProcess(rb));
+
+    // router.field not main sort field
+    rb = createResponseBuilderWQuery(queryComponent, createQueryReq(alias,
+        params("q", "*:*", "sort", "id asc, " + timeField + " asc")));
+    assertEquals(ResponseBuilder.STAGE_DONE, component.distributedProcess(rb));
+
+    // router.field is main sort field, component should kick in
+    rb = createResponseBuilderWQuery(queryComponent, createQueryReq(alias,
+        params("q", "*:*", "sort", timeField + " asc, id asc"))
+    );
+    assertEquals(ResponseBuilder.STAGE_GET_FIELDS, component.distributedProcess(rb));
+
+  }
+
+  private ResponseBuilder createResponseBuilderWQuery(QueryComponent queryComponent, SolrQueryRequest req) throws IOException {
+    ResponseBuilder rb = new ResponseBuilder(req, new SolrQueryResponse(), null);
+    rb.stage = ResponseBuilder.STAGE_START;
+    queryComponent.prepare(rb);
+    return rb;
   }
 
   private void assertUpdateResponse(UpdateResponse rsp) {
@@ -162,6 +242,10 @@ public class TRAQueryComponentTest extends SolrCloudTestCase {
   }
 
   private void createConfigSet(String configName) throws SolrServerException, IOException {
+    createConfigSet(configName, true);
+  }
+
+  private void createConfigSet(String configName, boolean deleteCollection) throws SolrServerException, IOException {
     // First create a configSet
     // Then we create a collection with the name of the eventual config.
     // We configure it, and ultimately delete the collection, leaving a modified config-set behind.
@@ -202,6 +286,11 @@ public class TRAQueryComponentTest extends SolrCloudTestCase {
             "  }" +
             "}").build()));
 
+    if (!deleteCollection) {
+      cluster.waitForActiveCollection(configName, 1, 1);
+      return;
+    }
+
     CollectionAdminRequest.deleteCollection(configName).process(solrClient);
     assertTrue(
         new ConfigSetAdminRequest.List().process(solrClient).getConfigSets()
@@ -224,6 +313,37 @@ public class TRAQueryComponentTest extends SolrCloudTestCase {
     return StrUtils.splitSmart(shards, ',').stream()
         .map(x -> x.substring(x.indexOf(SOLR_END_POINT) + SOLR_END_POINT.length()))
         .collect(Collectors.toList());
+  }
+
+  private static boolean collectionMatchesRegex(Collection<String> actual, String ... regexExp) {
+    for (String regex: regexExp) {
+      if (actual.stream().noneMatch(x -> x.matches(regex))) {
+        return false;
+      }
+    }
+    return true;
+  }
+
+  private static SolrQueryRequest createQueryReq(String collection, SolrParams params) throws IOException {
+    SolrQueryRequest reqMock = mock(LocalSolrQueryRequest.class, Mockito.RETURNS_DEEP_STUBS);
+    when(reqMock.getHttpSolrCall().getReq().getPathInfo()).thenReturn("/" + collection + "/select");
+    final SolrCore activeCore = getActiveCoreFromCluster();
+    when(reqMock.getCore()).then(AdditionalAnswers.delegatesTo(new LocalSolrQueryRequest(activeCore, params)));
+    when(reqMock.getSchema()).thenReturn(activeCore.getLatestSchema());
+    when(reqMock.getParams()).thenReturn(params);
+    return reqMock;
+  }
+
+  private static SolrCore getActiveCoreFromCluster() throws IOException {
+    Optional<JettySolrRunner> jettyRunner = cluster.getJettySolrRunners().stream()
+        .filter(x -> !x.getCoreContainer().getCores().isEmpty())
+        .findFirst();
+
+    if(jettyRunner.isPresent()) {
+      return  jettyRunner.get().getCoreContainer().getCores().iterator().next();
+    }
+
+    throw new IOException("could not find active core in Solr cluster");
   }
 
 }
